Fixes for re-enabling more MSVC level 4 warnings: media/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Possibly-uninitialized local variable

This also contains a small number of other cleanups/simplifications to nearby
code.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/377803004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281888 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 95ebe29ceb0e0dc10ae005a74c6f8eba0509868b

+++ media/cast/rtp_receiver/rtp_parser/rtp_parser_unittest.cc
@@ -39,7 +39,7 @@ class RtpParserTest : public ::testing::Test {
   void ExpectParsesPacket() {
     RtpCastHeader parsed_header;
     const uint8* payload = NULL;
-    size_t payload_size = -1;
+    size_t payload_size = static_cast<size_t>(-1);
     EXPECT_TRUE(rtp_parser_.ParsePacket(
         packet_, kPacketLength, &parsed_header, &payload, &payload_size));
 
@@ -63,7 +63,7 @@ class RtpParserTest : public ::testing::Test {
   void ExpectDoesNotParsePacket() {
     RtpCastHeader parsed_header;
     const uint8* payload = NULL;
-    size_t payload_size = -1;
+    size_t payload_size = static_cast<size_t>(-1);
     EXPECT_FALSE(rtp_parser_.ParsePacket(
         packet_, kPacketLength, &parsed_header, &payload, &payload_size));
   }
